+++ yasa/plotting.py
@@ -6,13 +6,14 @@ import numpy as np
 import pandas as pd
 import seaborn as sns
 import matplotlib.pyplot as plt
+import matplotlib.dates as mdates
 from lspopt import spectrogram_lspopt
 from matplotlib.colors import Normalize, ListedColormap
 
 __all__ = ["plot_hypnogram", "plot_spectrogram", "topoplot"]
 
 
-def plot_hypnogram(hypno, sf_hypno=1 / 30, lw=1.5, fill_color=None, ax=None):
+def plot_hypnogram(hyp, lw=1.5, highlight="REM", fill_color=None, ax=None):
     """
     Plot a hypnogram.
 
@@ -20,28 +21,14 @@ def plot_hypnogram(hypno, sf_hypno=1 / 30, lw=1.5, fill_color=None, ax=None):
 
     Parameters
     ----------
-    hypno : array_like
-        Sleep stage (hypnogram).
-
-        .. note::
-            The default hypnogram format in YASA is a 1D integer vector where:
-
-            * -2 = Unscored
-            * -1 = Artefact / Movement
-            * 0 = Wake
-            * 1 = N1 sleep
-            * 2 = N2 sleep
-            * 3 = N3 sleep
-            * 4 = REM sleep
-    sf_hypno : float
-        The current sampling frequency of the hypnogram, in Hz, e.g.
-
-        * 1/30 = 1 value per each 30 seconds of EEG data,
-        * 1 = 1 value per second of EEG data
+    hyp : :py:class:`yasa.Hypnogram`
+        A YASA hypnogram instance.
     lw : float
         Linewidth.
-    fill_color : str
-        Color to fill space above hypnogram line, optional.
+    highlight : str or None
+        Optional stage to highlight with alternate color.
+    fill_color : str or None
+        Optional color to fill space above hypnogram line.
     ax : :py:class:`matplotlib.axes.Axes`
         Axis on which to draw the plot, optional.
 
@@ -54,79 +41,101 @@ def plot_hypnogram(hypno, sf_hypno=1 / 30, lw=1.5, fill_color=None, ax=None):
     --------
     .. plot::
 
-        >>> import yasa
-        >>> import numpy as np
+        >>> from yasa import simulate_hypno
         >>> import matplotlib.pyplot as plt
-        >>> hypno = np.loadtxt("https://github.com/raphaelvallat/yasa/raw/master/notebooks/data_full_6hrs_100Hz_hypno_30s.txt")
-        >>> fig, ax = plt.subplots(1, 1, figsize=(7, 3), constrained_layout=True)
-        >>> ax = yasa.plot_hypnogram(hypno, fill_color="gainsboro", ax=ax)
+        >>> hyp = simulate_hypno(tib=300, seed=11)
+        >>> ax = hyp.plot_hypnogram()
+        >>> plt.tight_layout()
+
+    .. plot::
+
+        >>> from yasa import Hypnogram
+        >>> values = 4 * ["W", "N1", "N2", "N3", "REM"] + ["ART", "N2", "REM", "W", "UNS"]
+        >>> hyp = Hypnogram(values, freq="24min").upsample("30s")
+        >>> ax = hyp.plot_hypnogram(lw=2, fill_color="thistle")
+        >>> plt.tight_layout()
+
+    .. plot::
+
+        >>> fig, axes = plt.subplots(nrows=2, figsize=(6, 4), constrained_layout=True)
+        >>> hyp_a = simulate_hypno(n_stages=3, seed=99)
+        >>> hyp_b = simulate_hypno(n_stages=3, seed=99, start="2022-01-31 23:30:00")
+        >>> hyp_a.plot_hypnogram(lw=1, fill_color="whitesmoke", highlight=None, ax=axes[0])
+        >>> hyp_b.plot_hypnogram(lw=1, fill_color="whitesmoke", highlight=None, ax=axes[1])
     """
+    from yasa.hypno import Hypnogram  # Avoiding circular import
+
+    assert isinstance(hyp, Hypnogram), "`hypno` must be YASA Hypnogram."
+
+    # Work with a copy of the Hypnogram to not alter the original
+    hyp = hyp.copy()
+
     # Increase font size while preserving original
     old_fontsize = plt.rcParams["font.size"]
     plt.rcParams.update({"font.size": 18})
 
-    # Safety checks
-    assert isinstance(hypno, (np.ndarray, pd.Series, list)), "hypno must be an array."
-    hypno = np.asarray(hypno).astype(int)
-    assert (hypno >= -2).all() and (hypno <= 4).all(), "hypno values must be between -2 to 4."
-    assert hypno.ndim == 1, "hypno must be a 1D array."
-    assert isinstance(sf_hypno, (int, float)), "sf must be int or float."
+    ## Remap stages to be in desired y-axis order ##
+    # Start with default of all allowed labels
+    stage_order = hyp.labels.copy()
+    stages_present = hyp.hypno.unique()
+    # Remove Art/Uns from stage order, and place back individually at front to be higher on plot
+    art_str = stage_order.pop(stage_order.index("ART"))
+    uns_str = stage_order.pop(stage_order.index("UNS"))
+    if "ART" in stages_present:
+        stage_order.insert(0, art_str)
+    if "UNS" in stages_present:
+        stage_order.insert(0, uns_str)
+    # Put REM after WAKE if all 5 standard stages are allowed
+    if hyp.n_stages == 5:
+        stage_order.insert(stage_order.index("WAKE") + 1, stage_order.pop(stage_order.index("REM")))
+    # Reset the Hypnogram mapping so any future returns have this order
+    hyp.mapping = {stage: i for i, stage in enumerate(stage_order)}
 
-    bins = np.arange(hypno.size + 1) / (sf_hypno * 3600)
-    # Make sure that REM is displayed after Wake
-    hypno = pd.Series(hypno).map({-2: -2, -1: -1, 0: 0, 1: 2, 2: 3, 3: 4, 4: 1}).values
+    ## Extract values to plot ##
+    hypno = hyp.as_int()
+    # Reduce to breakpoints (where stages change) to avoid drawing individual lines for every epoch
+    hypno = hypno[hypno.shift().ne(hypno)]
+    # Extract x-values (bins) and y-values to plot
+    yvalues = hypno.to_numpy()
+    if hyp.start is not None:
+        final_bin_edge = pd.Timestamp(hyp.start) + pd.Timedelta(hyp.duration, unit="min")
+        bins = np.append(hypno.index.to_list(), final_bin_edge)
+        bins = [mdates.date2num(b) for b in bins]
+        xlabel = "Time"
+    else:
+        final_bin_edge = hyp.duration * 60
+        bins = np.append(hyp.timedelta[hypno.index].total_seconds(), final_bin_edge)
+        bins /= 60 if hyp.duration <= 90 else 3600
+        xlabel = "Time [mins]" if hyp.duration <= 90 else "Time [hrs]"
 
-    # Reduce data and bin edges to only moments of change in the hypnogram
-    # (to avoid drawing thousands of tiny individual lines when sf is high)
-    change_points = np.nonzero(np.ediff1d(hypno, to_end=1))
-    hypno = hypno[change_points]
-    bins = np.append(0, bins[change_points])
+    # Make mask to draw the highlighted stage
+    yvals_highlight = np.ma.masked_not_equal(yvalues, hyp.mapping.get(highlight))
 
-    # Make masks for REM and Artefact/Unscored
-    hypno_rem = np.ma.masked_not_equal(hypno, 1)
-    hypno_art_uns = np.ma.masked_greater(hypno, -1)
-    hypno_sleep = np.ma.masked_less(hypno, 0)
-
-    # Start the plot
+    # Open the figure
     if ax is None:
         ax = plt.gca()
 
     # Draw background filling
     if fill_color is not None:
-        ax.stairs(-1 * hypno.clip(0), bins, fill=True, color=fill_color, lw=0)
-    # Draw main hypnogram line
-    ax.stairs(-1 * hypno, bins, baseline=None, color="k", lw=lw)
-    # Draw REM and Artefact/Unscored highlighting
-    if 1 in hypno:
-        ax.hlines(-1 * hypno_rem, xmin=bins[:-1], xmax=bins[1:], color="red", lw=lw)
-    if -2 in hypno or -1 in hypno:
-        ax.hlines(-1 * hypno_art_uns, xmin=bins[:-1], xmax=bins[1:], color="grey", lw=lw)
-    # Determine y-axis labels and limits
-    if -2 in hypno and -1 in hypno:
-        # Both Unscored and Artefacts are present
-        ax.set_yticks([2, 1, 0, -1, -2, -3, -4])
-        ax.set_yticklabels(["Uns", "Art", "W", "R", "N1", "N2", "N3"])
-        ax.set_ylim(-4.5, 2.5)
-    elif -2 in hypno and -1 not in hypno:
-        # Only Unscored are present
-        ax.set_yticks([2, 0, -1, -2, -3, -4])
-        ax.set_yticklabels(["Uns", "W", "R", "N1", "N2", "N3"])
-        ax.set_ylim(-4.5, 2.5)
-    elif -2 not in hypno and -1 in hypno:
-        # Only Artefacts are present
-        ax.set_yticks([1, 0, -1, -2, -3, -4])
-        ax.set_yticklabels(["Art", "W", "R", "N1", "N2", "N3"])
-        ax.set_ylim(-4.5, 1.5)
-    else:
-        # No artefacts or Unscored
-        ax.set_yticks([0, -1, -2, -3, -4])
-        ax.set_yticklabels(["W", "R", "N1", "N2", "N3"])
-        ax.set_ylim(-4.5, 0.5)
-    ax.set_xlim(0, bins.max())
+        bline = hyp.mapping["WAKE"]  # len(stage_order) - 1 to fill from bottom
+        ax.stairs(yvalues.clip(bline), bins, baseline=bline, color=fill_color, fill=True, lw=0)
+    # Draw main hypnogram line, highlighted stage line, and Artefact/Unscored line
+    ax.stairs(yvalues, bins, baseline=None, color="black", lw=lw)
+    if not yvals_highlight.mask.all():
+        ax.hlines(yvals_highlight, xmin=bins[:-1], xmax=bins[1:], color="red", lw=lw)
+
+    # Aesthetics
+    ax.use_sticky_edges = False
+    ax.margins(x=0, y=1 / len(stage_order) / 2)  # 1/n_epochs/2 gives half-unit margins
+    ax.set_yticks(range(len(stage_order)))
+    ax.set_yticklabels(stage_order)
     ax.set_ylabel("Stage")
-    ax.set_xlabel("Time [hrs]")
-    ax.spines["right"].set_visible(False)
-    ax.spines["top"].set_visible(False)
+    ax.set_xlabel(xlabel)
+    ax.invert_yaxis()
+    ax.spines[["right", "top"]].set_visible(False)
+    if hyp.start is not None:
+        ax.xaxis.set_major_formatter(mdates.DateFormatter("%H:%M"))
+        ax.xaxis.set_major_locator(mdates.AutoDateLocator())
     # Revert font-size
     plt.rcParams.update({"font.size": old_fontsize})
     return ax
@@ -195,7 +204,7 @@ def plot_spectrogram(
     vmax : int or float
         The upper range of color scale. Overwrites ``trimperc``
     **kwargs : dict
-        Other arguments that are passed to :py:func:`yasa.plot_hypnogram`.
+        Other arguments that are passed to :py:meth:`yasa.Hypnogram.plot_hypnogram`.
 
     Returns
     -------
@@ -237,32 +246,33 @@ def plot_spectrogram(
         >>> hypno = yasa.hypno_upsample_to_data(hypno, 1/30, data, sf)
         >>> fig = yasa.plot_spectrogram(data, sf, hypno, cmap='Spectral_r')
     """
+    from yasa.hypno import Hypnogram, hypno_int_to_str  # Avoiding circular imports
+
     # Increase font size while preserving original
     old_fontsize = plt.rcParams["font.size"]
     plt.rcParams.update({"font.size": 18})
 
     # Safety checks
-    assert isinstance(data, np.ndarray), "Data must be a 1D NumPy array."
-    assert isinstance(sf, (int, float)), "sf must be int or float."
-    assert data.ndim == 1, "Data must be a 1D (single-channel) NumPy array."
-    assert isinstance(win_sec, (int, float)), "win_sec must be int or float."
-    assert isinstance(fmin, (int, float)), "fmin must be int or float."
-    assert isinstance(fmax, (int, float)), "fmax must be int or float."
-    assert fmin < fmax, "fmin must be strictly inferior to fmax."
-    assert fmax < sf / 2, "fmax must be less than Nyquist (sf / 2)."
-    assert isinstance(vmin, (int, float, type(None))), "vmin must be int, float, or None."
-    assert isinstance(vmax, (int, float, type(None))), "vmax must be int, float, or None."
+    assert isinstance(data, np.ndarray), "`data` must be a 1D NumPy array."
+    assert isinstance(sf, (int, float)), "`sf` must be int or float."
+    assert data.ndim == 1, "`data` must be a 1D (single-channel) NumPy array."
+    assert isinstance(win_sec, (int, float)), "`win_sec` must be int or float."
+    assert isinstance(fmin, (int, float)), "`fmin` must be int or float."
+    assert isinstance(fmax, (int, float)), "`fmax` must be int or float."
+    assert fmin < fmax, "`fmin` must be strictly inferior to `fmax`."
+    assert fmax < sf / 2, "`fmax` must be less than Nyquist (sf / 2)."
+    assert isinstance(vmin, (int, float, type(None))), "`vmin` must be int, float, or None."
+    assert isinstance(vmax, (int, float, type(None))), "`vmax` must be int, float, or None."
     if vmin is not None:
-        assert isinstance(vmax, (int, float)), "vmax must be int or float if vmin is provided"
+        assert isinstance(vmax, (int, float)), "`vmax` must be int or float if `vmin` is provided."
     if vmax is not None:
-        assert isinstance(vmin, (int, float)), "vmin must be int or float if vmax is provided"
+        assert isinstance(vmin, (int, float)), "`vmin` must be int or float if `vmax` is provided."
     if hypno is not None:
-        # Validate and handle hypnogram-related inputs
-        assert hypno.size == data.size, "Hypno must have the same sf as data."
+        assert hypno.size == data.size, "`hypno` must have the same number of samples as `data`."
 
     # Calculate multi-taper spectrogram
     nperseg = int(win_sec * sf)
-    assert data.size > 2 * nperseg, "Data length must be at least 2 * win_sec."
+    assert data.size > 2 * nperseg, "`data` length must be at least 2 * `win_sec`."
     f, t, Sxx = spectrogram_lspopt(data, sf, nperseg=nperseg, noverlap=0)
     Sxx = 10 * np.log10(Sxx)  # Convert uV^2 / Hz --> dB / Hz
 
@@ -294,10 +304,14 @@ def plot_spectrogram(
     ax1.set_xlabel("Time [hrs]")
 
     if hypno is not None:
+        # Convert sampling frequency to pandas timefrequency string (e.g., "30s")
+        freq_str = pd.tseries.frequencies.to_offset(pd.Timedelta(1 / sf, "S")).freqstr
+        # Create Hypnogram instance for plotting
+        hyp = Hypnogram(hypno_int_to_str(hypno), freq=freq_str)
         hypnoplot_kwargs = dict(lw=1.5, fill_color=None)
         hypnoplot_kwargs.update(kwargs)
         # Draw hypnogram
-        ax0 = plot_hypnogram(hypno, sf_hypno=sf, ax=ax0, **hypnoplot_kwargs)
+        ax0 = hyp.plot_hypnogram(ax=ax0, **hypnoplot_kwargs)
         ax0.xaxis.set_visible(False)
     else:
         # Add colorbar
@@ -406,13 +420,13 @@ def topoplot(
     plt.rcParams.update({"savefig.transparent": "True"})
 
     # Make sure we don't do any in-place modification
-    assert isinstance(data, pd.Series), "Data must be a Pandas Series"
+    assert isinstance(data, pd.Series), "`data` must be a Pandas Series"
     data = data.copy()
 
     # Add mask, if present
     if mask is not None:
-        assert isinstance(mask, pd.Series), "mask must be a Pandas Series"
-        assert mask.dtype.kind in "bi", "mask must be True/False or 0/1."
+        assert isinstance(mask, pd.Series), "`mask` must be a Pandas Series"
+        assert mask.dtype.kind in "bi", "`mask` must be True/False or 0/1."
     else:
         mask = pd.Series(1, index=data.index, name="mask")